refactor: rename util to utils on controller
