+++ src/adapters/controllers/role/create-role/create-role-controller.ts
@@ -7,7 +7,7 @@ import {
   serverError,
   badRequest,
   forbidden,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 
 export class CreateRoleController implements IController {
   constructor(
