+++ src/adapters/controllers/role/delete-role-by-id/delete-role-by-id-controller.ts
@@ -6,7 +6,7 @@ import {
   unauthorized,
   ok,
   badRequest,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 
 export class DeleteRoleByIdController implements IController {
   constructor(
