+++ src/adapters/controllers/user-has-role/add-role-to-user/add-role-to-user-controller.ts
@@ -6,7 +6,7 @@ import {
   serverError,
   badRequest,
   forbidden,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 
 export class AddRoleToUserController implements IController {
   constructor(
