+++ src/adapters/controllers/user-has-role/remove-role-from-user/remove-role-from-user-controller.ts
@@ -6,7 +6,7 @@ import {
   serverError,
   badRequest,
   forbidden,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 
 export class RemoveRoleFromUserController implements IController {
   constructor(
