+++ test/adapters/controllers/user/read-all-users/read-all-users-controller.spec.ts
@@ -2,7 +2,7 @@ import { getUserCriticalData } from '@/adapters/controllers/user/utils';
 import { makeUserRepository } from '@/test/stubs';
 import { ReadAllUsersUseCase } from '@/use-cases/user/read-all-users';
 import { ServerError } from '@/adapters/errors';
-import { serverError } from '@/adapters/util/http';
+import { serverError } from '@/adapters/utils/http';
 import { UserBuilder } from '@/test/builders/user-builder';
 import { ReadAllUsersController } from '@/adapters/controllers/user/read-all-users';
