+++ test/adapters/controllers/verification-token/send-verification-token-to-the-user/send-verification-token-to-the-user-controller.spec.ts
@@ -12,7 +12,7 @@ import {
   forbidden,
   ok,
   unauthorized,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 import {
   makeFakeRequest,
   makeEncrypter,
