+++ test/adapters/controllers/verification-token/verify-token/verify-token-controller.spec.ts
@@ -9,7 +9,7 @@ import {
   forbidden,
   ok,
   unauthorized,
-} from '@/adapters/util/http';
+} from '@/adapters/utils/http';
 import {
   makeValidation,
   makeVerificationTokenRepository,
